feat: Overhaul extension with major refactor and optimizations
This commit introduces a comprehensive overhaul of the extension, including significant performance optimizations, modernization of the toolchain, critical bug fixes, and a complete code refactoring.

Key Changes:

- **Performance:**
    - Changed activation events from `*` to specific commands (`onCommand:extension.*`) to prevent unnecessary activation on VS Code startup, improving performance.

- **Dependency & Security Management:**
    - Upgraded core dependency `@vitalets/google-translate-api` to v9.
    - Removed unused packages `google-translate-open-api` and `tunnel`.
    - Modernized the testing framework by replacing the deprecated `vscode` package with `@vscode/test-electron`, `mocha`, and `glob`. This resolved multiple security vulnerabilities.
    - Installed `http-proxy-agent` to support the new translation API.

- **Bug Fixes:**
    - Fixed a critical runtime error (`gti is not a function`) caused by breaking changes in `@vitalets/google-translate-api` v9.
    - Refactored translation and proxy logic to be compatible with the new API.

- **Refactoring:**
    - Decomposed the monolithic `extension.js` into smaller, single-responsibility modules under a new `src/` directory (`commands.js`, `translate.js`, `utils.js`).
    - This significantly improves code readability, separation of concerns, and maintainability.

- **Testing:**
    - Implemented a robust integration test suite that verifies core translation functionality.
    - The new tests would have caught the critical bugs that were fixed, ensuring higher stability for future releases.

- **Feature Removal:**
    - Removed the "Tooltip Hover Translation" feature, including the language server and all related UI and dependencies, to simplify the codebase.
