--- server/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "comment-translate-server",
-  "version": "1.0.0",
-  "description": "vscode comment translate server",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "dependencies": {
-    "@vitalets/google-translate-api": "^8.0.0",
-    "google-translate-open-api": "^1.3.7",
-    "humanize-string": "^3.0.0",
-    "tunnel": "0.0.6",
-    "onigasm": "^2.2.5",
-    "request": "^2.88.2",
-    "vscode-languageserver": "^9.0.1",
-    "vscode-textmate": "^9.0.0"
-  },
-  "license": "ISC",
-  "devDependencies": {
-    "@types/node": "^20.11.24",
-    "@types/request": "^2.48.12"
-  }
-}
