--- server/tsconfig.json
@@ -1,17 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es6",
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"sourceMap": true,
-		"outDir": "out",
-		"rootDir": "src",
-		"lib": [ "es6" ]
-	},
-	"include": [
-		"src"
-	],
-	"exclude": [
-		"node_modules"
-	]
-}
